+++ src/CosmosDB/CosmosDB/Restore/GetAzCosmosDBTableRestorableTable.cs
@@ -1,4 +1,4 @@
-﻿// ----------------------------------------------------------------------------------
+// ----------------------------------------------------------------------------------
 //
 // Copyright Microsoft Corporation
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -21,8 +21,8 @@ using Microsoft.Azure.Management.CosmosDB.Models;
 
 namespace Microsoft.Azure.Commands.CosmosDB
 {
-    [Cmdlet(VerbsCommon.Get, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "CosmosDBMongoDBRestorableCollection", DefaultParameterSetName = NameParameterSet), OutputType(typeof(PSRestorableMongodbCollectionGetResult))]
-    public class GetAzCosmosDBMongoDBRestorableCollection : AzureCosmosDBCmdletBase
+    [Cmdlet(VerbsCommon.Get, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "CosmosDBTableRestorableTable", DefaultParameterSetName = NameParameterSet), OutputType(typeof(PSRestorableTableGetResult))]
+    public class GetAzCosmosDBTableRestorableTable : AzureCosmosDBCmdletBase
     {
         [Parameter(Mandatory = true, ParameterSetName = NameParameterSet, HelpMessage = Constants.LocationNameHelpMessage)]
         [ValidateNotNullOrEmpty]
@@ -32,19 +32,15 @@ namespace Microsoft.Azure.Commands.CosmosDB
         [ValidateNotNullOrEmpty]
         public string DatabaseAccountInstanceId { get; set; }
 
-        [Parameter(Mandatory = true, ParameterSetName = NameParameterSet, HelpMessage = Constants.DatabaseResourceIdHelpMessage)]
-        [ValidateNotNullOrEmpty]
-        public string DatabaseRId { get; set; }
-
-        [Parameter(Mandatory = true, ValueFromPipeline = true, ParameterSetName = ParentObjectParameterSet, HelpMessage = Constants.RestorableMongoDBDatabaseObjectHelpMessage)]
+        [Parameter(Mandatory = true, ValueFromPipeline = true, ParameterSetName = ParentObjectParameterSet, HelpMessage = Constants.RestorableTableObjectHelpMessage)]
         [ValidateNotNull]
-        public PSRestorableSqlDatabaseGetResult InputObject { get; set; }
+        public PSRestorableTableGetResult InputObject { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = NameParameterSet, HelpMessage = Constants.RestorableMongoDBCollectionsFeedStartTimeHelpMessage)]
+        [Parameter(Mandatory = false, ParameterSetName = NameParameterSet, HelpMessage = Constants.RestorableTablesFeedStartTimeHelpMessage)]
         [ValidateNotNullOrEmpty]
         public string StartTime { get; set; }
 
-        [Parameter(Mandatory = false, ParameterSetName = NameParameterSet, HelpMessage = Constants.RestorableMongoDBCollectionsFeedEndTimeHelpMessage)]
+        [Parameter(Mandatory = false, ParameterSetName = NameParameterSet, HelpMessage = Constants.RestorableTablesFeedEndTimeHelpMessage)]
         [ValidateNotNullOrEmpty]
         public string EndTime { get; set; }
 
@@ -52,17 +48,16 @@ namespace Microsoft.Azure.Commands.CosmosDB
         {
             if (ParameterSetName.Equals(ParentObjectParameterSet, StringComparison.Ordinal))
             {
-                // id is in the format: /subscriptions/<subscriptionId>/providers/Microsoft.DocumentDB/locations/<location>/restorableDatabaseAccounts/<DatabaseAccountInstanceId>/restorableMongoDBDatabases/<Id>
+                // id is in the format: /subscriptions/<subscriptionId>/providers/Microsoft.DocumentDB/locations/<location>/restorableDatabaseAccounts/<DatabaseAccountInstanceId>/<Id>
                 string[] idComponents = InputObject.Id.Split('/');
                 Location = HttpUtility.UrlDecode(idComponents[6]);
                 DatabaseAccountInstanceId = idComponents[8];
-                DatabaseRId = InputObject.OwnerResourceId;
             }
 
-            IEnumerable restorableMongoDBCollections = CosmosDBManagementClient.RestorableMongodbCollections.ListWithHttpMessagesAsync(Location, DatabaseAccountInstanceId, DatabaseRId, StartTime, EndTime).GetAwaiter().GetResult().Body;
-            foreach (RestorableMongodbCollectionGetResult restorableMongoDBCollection in restorableMongoDBCollections)
+            IEnumerable restorableTables = CosmosDBManagementClient.RestorableTables.ListWithHttpMessagesAsync(Location, DatabaseAccountInstanceId, StartTime, EndTime).GetAwaiter().GetResult().Body;
+            foreach (RestorableTableGetResult restorableTable in restorableTables)
             {
-                WriteObject(new PSRestorableMongodbCollectionGetResult(restorableMongoDBCollection));
+                WriteObject(new PSRestorableTableGetResult(restorableTable));
             }
         }
     }